Extract nick, myroomnick, myroomavatar, myavatar to profile
